test_requirements.txt
pytest==2.9.1
pytest-django==2.9.1
mock==2.0.0
